chores: move unused files to legacy directory
