+++ legacy/BoardIntroduce.tsx
