+++ legacy/debounced-query.ts
@@ -1,5 +1,5 @@
-import { useSearchStore } from "..";
-import { useDebounce } from "../../../shared/model/useDebounce";
+import { useSearchStore } from "../src/features/search";
+import { useDebounce } from "../src/shared/model/useDebounce";
 
 export const useDebouncedQuery = () => {
   const query = useSearchStore((state) => state.query);
