+++ legacy/delete-post.ts
@@ -1,7 +1,7 @@
 import { useMutation, useQueryClient } from "@tanstack/react-query";
-import instance from "../../../shared/api/instance";
-import { POST_DETAILS_QUERY_KEY } from "../../../entities/post/api/post-details.query";
-import { POST_LIST_QUERY_KEY } from "../../../entities/post/api/post-list.query";
+import instance from "../src/shared/api/instance";
+import { POST_DETAILS_QUERY_KEY } from "./post-details.query";
+import { POST_LIST_QUERY_KEY } from "./post-list.query";
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 export const deletePost = async (id: string): Promise<any> => {
