+++ legacy/edit-post.ts
@@ -1,7 +1,7 @@
 import { useMutation, useQueryClient } from "@tanstack/react-query";
-import instance from "../../../shared/api/instance";
-import { Post } from "../../../shared/types/post";
-import { POST_DETAILS_QUERY_KEY } from "../../../entities/post/api/post-details.query";
+import instance from "../src/shared/api/instance";
+import { Post } from "./post";
+import { POST_DETAILS_QUERY_KEY } from "./post-details.query";
 
 const editPost = async ({ id, post }: { id: string; post: Post }) => {
   const response = await instance.patch(`posts/${id}`, post);
