+++ legacy/submit-post.ts
@@ -1,7 +1,9 @@
+"use client";
+
 import { useMutation, useQueryClient } from "@tanstack/react-query";
-import { Post } from "../../../shared/types/post";
-import instance from "../../../shared/api/instance";
-import { POST_LIST_QUERY_KEY } from "../../../entities/post/api/post-list.query";
+import { Post } from "./post";
+import instance from "../src/shared/api/instance";
+import { POST_LIST_QUERY_KEY } from "./post-list.query";
 
 const submitPost = async (newPost: Post) => instance.post("/posts", newPost);
