+++ src/entities/post/index.ts
@@ -1,7 +1,1 @@
-export { PostCard } from "./ui/PostCard";
-
-// export { LikeStat as LikeStat } from "./ui/LikeStat";
-
-// export { useGetDetailsQuery as useGetPostDetailsQuery } from "./api/post-details.query";
-// export { useGetListQuery as useGetPostListQuery } from "./api/post-list.query";
-// export { useGetPage as useGetPostPage } from "./api/post-list.query";
+export { Card as PostCard } from "./ui/Card";
