--- src/features/post/index.ts
@@ -1,7 +1,0 @@
-export { LikeButton as PostLikeButton } from "./ui/LikeButton";
-export { CreateNewButton as CreateNewPostButton } from "./ui/CreateNewButton";
-
-export { useSubmit as useSubmitPost } from "./api/submit-post";
-export { useDelete as useDeletePost } from "./api/delete-post";
-export { useEdit as useEditPost } from "./api/edit-post";
-export { useLike as useLikePost } from "./api/like-post";
