--- src/features/search/index.ts
@@ -1,2 +1,0 @@
-export { useMatchingList as useMatchingPostList } from "./model/matching-post-list";
-export { useStore as useSearchStore } from "./model/search-store";
